MVC NickList using FXML
