+++ src/main/kotlin/com/dmdirc/WindowUI.kt
@@ -9,6 +9,8 @@ import com.dmdirc.Style.CustomStyle
 import com.dmdirc.edgar.Edgar.tr
 import com.dmdirc.ktirc.events.ChannelMembershipAdjustment
 import com.dmdirc.ktirc.events.IrcEvent
+import com.dmdirc.ui.nicklist.NickListController
+import com.dmdirc.ui.nicklist.NickListModel
 import com.uchuhimo.konf.Item
 import javafx.application.HostServices
 import javafx.beans.Observable
@@ -18,7 +20,6 @@ import javafx.beans.property.SimpleObjectProperty
 import javafx.beans.property.SimpleStringProperty
 import javafx.collections.ListChangeListener
 import javafx.geometry.Orientation.VERTICAL
-import javafx.scene.control.ListView
 import javafx.scene.control.ScrollBar
 import javafx.scene.control.TextArea
 import javafx.scene.control.TextField
@@ -152,11 +153,7 @@ class WindowUI(model: WindowModel, hostServices: HostServices, imageLoader: (Str
                 }
             }
             if (!model.isConnection) {
-                right = ListView<String>(model.nickList.users).apply {
-                    isFocusTraversable = false
-                    styleClass.add("nick-list")
-                    prefWidth = 148.0
-                }
+                right = NickListController(model.nickList).createUi()
             }
             bottom = inputField
             AnchorPane.setTopAnchor(this, 0.0)
