+++ src/main/kotlin/com/dmdirc/ui/nicklist/NickListContract.kt
@@ -1,0 +1,15 @@
+package com.dmdirc.ui.nicklist
+
+import javafx.collections.ObservableList
+import javafx.scene.Node
+
+object NickListContract {
+
+    interface Model {
+        val users: ObservableList<String>
+    }
+
+    interface Controller {
+        fun createUi(): Node
+    }
+}
